Changing Motor method names because redefinition of close
The gpiozero library defines a close method for CompositeDevice.
Motor class inherits from CompositeDevice, so the close method is
already used in a very different way (free resources, mainly).

This name change will fix the extrange behavior of the unit tests
that apparently takes 6 seconds to run even when the time.sleep function
was already mocked